market-starter/requirements.txt
# ── Core ───────────────────────────────────────────────
streamlit==1.54.0
sqlalchemy==2.0.48
alembic==1.18.4
psycopg[binary]==3.3.2
python-dotenv==1.1.0
pandas==2.2.3

# ── Dev / CI ───────────────────────────────────────────
ruff==0.15.4
black==26.1.0
pytest==9.0.2
